requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data Formats
pyarrow>=12.0.0

# LLM Integration
openai>=1.0.0
anthropic>=0.20.0

# CLI Framework
click>=8.0.0

# Machine Learning & Analysis
scikit-learn>=1.3.0

# Data Validation
pydantic>=2.0.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Progress Bars
tqdm>=4.65.0

# Async Support (optional)
aiohttp>=3.8.0

# Development Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# Type Stubs
types-requests>=2.31.0
